Fix for Enhancement HZN-565 - Part II
Added simple JUnit tests to verify the configuration file and the
content of the translated sendmail object.

The AbstractNorthbounder has to be modified in order to use ${descr} as
a valid placeholder.

Also, there is no need to keep a Map with the Mappings based on the
alarmId, as I think it is impossible to process the exact same alarm
twice by the same northbounder implementation, besides the fact that it
doesn't save resources by creating the mapping for each received alarm.